修改版本和模块名称

+++ publish-jar.sh
@@ -1,7 +1,7 @@
 #!/bin/sh
 ## sed -i 's/\r$//' publish-jar.sh & chmod 777 publish-jar.sh & ./publish-jar.sh
 BUILD_ID=DONTKILLME #后台执行
-arr=("gulimall-product" "gulimall-coupon" "gulimall-member" "gulimall-order" "gulimall-ware" "gulimall_gateway" "renren-fast")
+arr=("gulimall-product" "gulimall_gateway" "renren-fast")
 # shellcheck disable=SC2068
 for k in ${arr[@]}; do
   houzhui="-0.0.1-SNAPSHOT.jar"
@@ -18,16 +18,35 @@ for k in ${arr[@]}; do
     mkdir -p /zkl/project/gulimall/log
   fi
   #copy 之前先删除
-  rm -rf /zkl/project/gulimall/
-  #copy jar 到启动目录
-  cp -r /root/.jenkins/workspace/gulimall/"${k}"/target/"${jar_name}" /zkl/project/gulimall/"${jar_name}"
+  rm -rf /zkl/project/gulimall/"${jar_name}"
+  rm -rf /zkl/project/gulimall/log/"${k}".log
 
+  # 如果文件存在那么进行复制，否则睡五秒之后再复制
+  if [  -f /root/.jenkins/workspace/gulimall/"${k}"/target/"${jar_name}" ]; then
+     #copy jar 到启动目录
+    cp -r /root/.jenkins/workspace/gulimall/"${k}"/target/"${jar_name}" /zkl/project/gulimall/"${jar_name}"
+  else
+    sleep 4s
+    cp -r /root/.jenkins/workspace/gulimall/"${k}"/target/"${jar_name}" /zkl/project/gulimall/"${jar_name}"
+  fi
+
+  #如果没有文件等待五秒后再看是否复制完成
   if [ ! -f /zkl/project/gulimall/"${jar_name}" ]; then
-    sleep 5s
+    sleep 10s
   fi
+  #证明没有复制成功，那么continue
+  if [ ! -f /zkl/project/gulimall/"${jar_name}" ]; then
+    continue
+  fi
+  #日志文件不存在，则创建一个日志文件
   if [ ! -f /zkl/project/gulimall/log/"${k}".log ]; then
-    touch /zkl/project/gulimall/log/"${k}".log
+    sudo touch /zkl/project/gulimall/log/"${k}".log
   fi
-  nohup java -jar /zkl/project/gulimall/"${jar_name}" >/zkl/project/gulimall/log/"${k}".log 2>&1 &
-  echo "${k} 已启动！"
+  if [ -f /zkl/project/gulimall/"${jar_name}" ]; then
+      nohup java -jar /zkl/project/gulimall/"${jar_name}" >/zkl/project/gulimall/log/"${k}".log 2>&1 &
+      echo "${k} 已启动！"
+  else
+     echo "${k} 启动失败！"
+  fi
+
 done
